+++ mp4v2 wrapper/MP42File.m
@@ -85,12 +85,12 @@
 
 - (void)removeTracksAtIndexes:(NSIndexSet *)indexes
 {
-  unsigned index = [indexes firstIndex];
-  while (index!=NSNotFound) {    
+  NSUInteger index = [indexes firstIndex];
+  while (index != NSNotFound) {    
     MP42Track *track = [tracks objectAtIndex:index];
     if (track.muxed)
       [tracksToBeDeleted addObject:track];    
-    index=[indexes indexGreaterThanIndex:index];
+    index = [indexes indexGreaterThanIndex:index];
   }
   
   [tracks removeObjectsAtIndexes:indexes];
